fix: apply the correct architecture and remove unused files
