--- tests/models/dataset/conftest.py
@@ -1,3 +1,0 @@
-pytest_plugins = [
-    "tests.models.dataset.fixtures.dataset_organizer_fixtures",
-]
