+++ tests/models/dataset/fixtures/dataset_collection_fixtures.py
@@ -4,8 +4,8 @@ import pytest
 from picsellia.types.enums import InferenceType
 
 from src.models.dataset.dataset_split_name import DatasetSplitName
-from src.steps.data_extraction.utils.dataset_collection import DatasetCollection
-from tests.steps.data_extraction.utils.conftest import DatasetTestMetadata
+from src.models.dataset.dataset_collection import DatasetCollection
+from tests.steps.fixtures.dataset_version_fixtures import DatasetTestMetadata
 
 
 @pytest.fixture
